DROTH-3578 moved getChangedSegmentMeasures to LaneService
