+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/service/lane/LaneService.scala
@@ -16,7 +16,7 @@ import fi.liikennevirasto.digiroad2.service.{RoadAddressForLink, RoadAddressServ
 import fi.liikennevirasto.digiroad2.util.ChangeLanesAccordingToVvhChanges.updateChangeSet
 import fi.liikennevirasto.digiroad2.util.LaneUtils.{persistedHistoryLanesToTwoDigitLaneCode, persistedLanesTwoDigitLaneCode}
 import fi.liikennevirasto.digiroad2.util.{LaneUtils, LinearAssetUtils, LogUtils, PolygonTools}
-import fi.liikennevirasto.digiroad2.{DigiroadEventBus, GeometryUtils}
+import fi.liikennevirasto.digiroad2.{DigiroadEventBus, GeometryUtils, RoadAddress}
 import org.joda.time.DateTime
 import org.slf4j.LoggerFactory
 
@@ -24,6 +24,7 @@ import java.security.InvalidParameterException
 
 
 case class LaneChange(lane: PersistedLane, oldLane: Option[PersistedLane], changeType: LaneChangeType, roadLink: Option[RoadLink], historyEventOrderNumber: Option[Int])
+case class ChangedSegment(startMeasure: Double, startAddrM: Int, endMeasure: Double, endAddrM: Int, segmentChangeType: LaneChangeType)
 
 class LaneService(roadLinkServiceImpl: RoadLinkService, eventBusImpl: DigiroadEventBus, roadAddressServiceImpl: RoadAddressService) extends LaneOperations {
   override def roadLinkService: RoadLinkService = roadLinkServiceImpl
@@ -544,6 +545,143 @@ trait LaneOperations {
       case _ => sortedLanesChanged
     }
 
+  }
+
+
+  /**
+    *
+    * @param laneChange laneChange for shortened or lengthened lane
+    * @return Optional M-values and road address values for segments. 1st element of tuple is for digitizing
+    *         direction start segment, 2nd is for end segment. Returns None for segment in question if
+    *         that end of lane was not changed.
+    */
+  def getChangedSegmentMeasures(laneChange: LaneChange): (Option[ChangedSegment], Option[ChangedSegment]) = {
+    val roadLink = laneChange.roadLink.get
+    val lane = laneChange.lane
+    val oldLane = laneChange.oldLane.get
+
+    val roadLinkStartAddr = roadLink.attributes("START_ADDR").toString.toInt
+    val roadLinkEndAddr = roadLink.attributes("END_ADDR").toString.toInt
+    val roadAddressSideCode = SideCode.apply(roadLink.attributes("SIDECODE").toString.toInt)
+
+    val (laneStartAddrM, laneEndAddrM) = RoadAddress.getAddressMValuesForCutAssets(roadLink.length, roadAddressSideCode,
+      roadLinkStartAddr, roadLinkEndAddr, lane.startMeasure, lane.endMeasure)
+
+    val (oldLaneStartAddrM, oldLaneEndAddrM) = RoadAddress.getAddressMValuesForCutAssets(roadLink.length, roadAddressSideCode,
+      roadLinkStartAddr, roadLinkEndAddr, oldLane.startMeasure, oldLane.endMeasure)
+
+
+    val cutFromStart = lane.startMeasure > oldLane.startMeasure
+    val cutFromEnd = lane.endMeasure < oldLane.endMeasure
+    val lenghtenedFromStart = lane.startMeasure < oldLane.startMeasure
+    val lenghtenedFromEnd = lane.endMeasure > oldLane.endMeasure
+
+    (cutFromStart, cutFromEnd, lenghtenedFromStart, lenghtenedFromEnd) match {
+      // Lane cut from digitizing direction start
+      case (true, false, false, false) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, oldLaneStartAddrM, lane.startMeasure, laneStartAddrM, LaneChangeType.Expired))
+            val endSegmentChange = None
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, laneEndAddrM, lane.startMeasure, oldLaneEndAddrM, LaneChangeType.Expired))
+            val endSegmentChange = None
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane cut from digitizing direction end
+      case (false, true, false, false) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = None
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneEndAddrM, oldLane.endMeasure, oldLaneEndAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = None
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneStartAddrM, oldLane.endMeasure, oldLaneStartAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane cut from both ends
+      case (true, true, false, false) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, oldLaneStartAddrM, lane.startMeasure, laneStartAddrM, LaneChangeType.Expired))
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneEndAddrM, oldLane.endMeasure, oldLaneEndAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, oldLaneEndAddrM, lane.startMeasure, laneEndAddrM, LaneChangeType.Expired))
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneStartAddrM, oldLane.endMeasure, oldLaneStartAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+
+        }
+
+      // Lane lengthened from digitizing direction start
+      case (false, false, true, false) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, laneStartAddrM, oldLane.startMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            val endSegmentChange = None
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, oldLaneEndAddrM, oldLane.startMeasure, laneEndAddrM, LaneChangeType.Add))
+            val endSegmentChange = None
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane lengthened from digitizing direction end
+      case (false, false, false, true) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = None
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, oldLaneEndAddrM, lane.endMeasure, laneEndAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = None
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, laneStartAddrM, lane.endMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane lengthened from both ends
+      case (false, false, true, true) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, laneStartAddrM, oldLane.startMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, oldLaneEndAddrM, lane.endMeasure, laneEndAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, oldLaneEndAddrM, oldLane.startMeasure, laneEndAddrM, LaneChangeType.Add))
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, laneStartAddrM, lane.endMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane cut from digitizing direction start and lengthened from end
+      case (true, false, false, true) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, oldLaneStartAddrM, lane.startMeasure, laneStartAddrM, LaneChangeType.Expired))
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, oldLaneEndAddrM, lane.endMeasure, laneEndAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(oldLane.startMeasure, oldLaneEndAddrM, lane.startMeasure, laneEndAddrM, LaneChangeType.Expired))
+            val endSegmentChange = Some(ChangedSegment(oldLane.endMeasure, laneStartAddrM, lane.endMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            (startSegmentChange, endSegmentChange)
+        }
+
+      // Lane cut from digitizing direction end and lengthened from start
+      case (false, true, true, false) =>
+        roadAddressSideCode match {
+          case SideCode.TowardsDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, laneStartAddrM, oldLane.startMeasure, oldLaneStartAddrM, LaneChangeType.Add))
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneEndAddrM, oldLane.endMeasure, oldLaneEndAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+          case SideCode.AgainstDigitizing =>
+            val startSegmentChange = Some(ChangedSegment(lane.startMeasure, oldLaneEndAddrM, oldLane.startMeasure, laneEndAddrM, LaneChangeType.Add))
+            val endSegmentChange = Some(ChangedSegment(lane.endMeasure, laneStartAddrM, oldLane.endMeasure, oldLaneStartAddrM, LaneChangeType.Expired))
+            (startSegmentChange, endSegmentChange)
+        }
+    }
   }
 
   def persistedLanesToPwLanesWithAddress(lanes: Seq[PersistedLane], roadLinks: Seq[RoadLink]): Seq[PieceWiseLane] = {
